Revert "refactor: optimize PR detail API and improve pullRequests.ts readabil…"
This reverts commit b82594c4afdf9f904b7a9f584f8188218dd84bd2.
