+++ backend/types/pullRequests.ts
@@ -46,8 +46,6 @@ export type GhPRCommit = {
 export type GhReviewInlineComment = {
   id: number;
   node_id: string;
-  /** Numeric ID of the review this comment belongs to */
-  pull_request_review_id: number;
   /** ID of the comment this is a reply to. Absent if this is the first comment in the thread */
   in_reply_to_id?: number;
   user: {
